Модели выведены в отдельные файлы Добавлена фабрика вопросов, выведенная в отдельный файл
