+++ MovieQuiz/Models/QuizResultViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultViewModel.swift
+//  MovieQuiz
+//
+//  Created by Egor Partenko on 30.10.2023.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
