+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -1,31 +1,15 @@
 import UIKit
 
-struct QuizQuestion {
-       let image: String
-       let text: String
-       let correctAnswer: Bool
-   }
-   
-   struct QuizStepViewModel {
-       let image: UIImage
-       let question: String
-       let questionNumber: String
-   }
-   
-   struct QuizResultsViewModel {
-       let title: String
-       let text: String
-       let buttonText: String
-   }
-
 final class MovieQuizViewController: UIViewController {
     // MARK: - Lifecycle
     override func viewDidLoad() {
            super.viewDidLoad()
-           let currentQuestion = questions[currentQuestionIndex]
-           let viewModel = convert(model: currentQuestion)
-           self.show(quiz: viewModel)
-       }
+        if let firstQuestion = questionFactory.requestNextQuestion() {
+            currentQuestion = firstQuestion
+            let viewModel = convert(model: firstQuestion)
+            show(quiz: viewModel)
+        }
+    }
     
     @IBAction func yesButtonClicked(_ sender: UIButton) {
         answerGived(answer: true)
@@ -45,49 +29,9 @@ final class MovieQuizViewController: UIViewController {
     
     private var currentQuestionIndex = 0
     private var correctAnswers = 0
-    
-    private let questions: [QuizQuestion] = [
-            QuizQuestion(
-                image: "The Godfather",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Dark Knight",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Kill Bill",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Avengers",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Deadpool",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "The Green Knight",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: true),
-            QuizQuestion(
-                image: "Old",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "The Ice Age Adventures of Buck Wild",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "Tesla",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false),
-            QuizQuestion(
-                image: "Vivarium",
-                text: "Рейтинг этого фильма больше чем 6?",
-                correctAnswer: false)
-        ]
+    private let questionsAmount = 10
+    private var questionFactory: QuestionFactory = QuestionFactory()
+    private var currentQuestion: QuizQuestion?
     
     private func showAnswerResult(isCorrect: Bool){
         
@@ -112,7 +56,7 @@ final class MovieQuizViewController: UIViewController {
            let questionStep = QuizStepViewModel(
                image: UIImage(named: model.image) ?? UIImage() ,
                question: model.text,
-               questionNumber: "\(currentQuestionIndex + 1)/\(questions.count)")
+               questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)")
            return questionStep
        }
     
@@ -133,10 +77,12 @@ final class MovieQuizViewController: UIViewController {
             self.currentQuestionIndex = 0
             self.correctAnswers = 0
             
-            let currentQuestion = self.questions[self.currentQuestionIndex]
-            let viewModel = self.convert(model: currentQuestion)
-            self.show(quiz: viewModel)
+            if let firstQuestion = self.questionFactory.requestNextQuestion() {
+                self.currentQuestion = firstQuestion
+                let viewModel = self.convert(model: firstQuestion)
+                self.show(quiz: viewModel)
             }
+        }
         alert.addAction(action)
         
         self.present(alert, animated: true, completion: nil)
@@ -144,8 +90,11 @@ final class MovieQuizViewController: UIViewController {
     }
     
     private func showNextQuestionOrResult() {
-        if currentQuestionIndex == questions.count - 1 {
-            let text = "Ваш результат \(correctAnswers)/10"
+        if currentQuestionIndex == questionsAmount - 1 {
+            let text = correctAnswers == questionsAmount ?
+            "Поздравляем, вы ответили на 10 из 10!" :
+            "Ваш результат \(correctAnswers)/10"
+            
             let viewModelResult = QuizResultsViewModel(
                 title: "Игра окончена!",
                 text: text,
@@ -156,14 +105,18 @@ final class MovieQuizViewController: UIViewController {
             
             currentQuestionIndex += 1
             
-            let nextQuestion = questions[currentQuestionIndex]
-            let viewModel = convert(model: nextQuestion)
-            show(quiz: viewModel)
+            if let nextQuestion = questionFactory.requestNextQuestion() {
+                currentQuestion = nextQuestion
+                let viewModel = convert(model: nextQuestion)
+                show(quiz: viewModel)
+            }
         }
     }
     
     private func answerGived(answer: Bool) {
-        let currentQuestion = questions[currentQuestionIndex]
+        guard let currentQuestion = currentQuestion else {
+            return
+        }
         let givenAnswer = answer
         showAnswerResult(isCorrect: givenAnswer == currentQuestion.correctAnswer)
     }
